Trait aws_hyper::conn::HttpService

pub trait HttpService: Send + Sync {
    fn poll_ready(
        &mut self, 
        cx: &mut Context<'_>
    ) -> Poll<Result<(), Box<dyn Error + Send + Sync>>>;

    fn call(
        &mut self, 
        req: Request<SdkBody>
    ) -> Pin<Box<dyn Future<Output = Result<Response<SdkBody>, Box<dyn Error + Send + Sync>>> + Send>>;

    fn clone_box(&self) -> Box<dyn HttpService>;
}

Required methods

fn poll_ready(
&mut self,
cx: &mut Context<'_>
) -> Poll<Result<(), Box<dyn Error + Send + Sync>>>

Return whether this service is ready to accept a request

See Service::poll_ready

fn call(
&mut self,
req: Request<SdkBody>
) -> Pin<Box<dyn Future<Output = Result<Response<SdkBody>, Box<dyn Error + Send + Sync>>> + Send>>

Call this service and return a response

See Service::call

fn clone_box(&self) -> Box<dyn HttpService>

Return a Boxed-clone of this service

aws_hyper::Client will clone the inner service for each request so this should be a cheap clone operation.