Struct aws_hyper::Client

pub struct Client<S> { /* fields omitted */ }

AWS Service Client

Hyper-based AWS Service Client. Most customers will want to construct a client with Client::https(). For testing & other more advanced use cases, a custom connector may be used via Client::new(connector).

The internal connector must implement the following trait bound to be used to dispatch requests:

   S: Service<http::Request<SdkBody>, Response = http::Response<hyper::Body>>
       + Send
       + Clone
       + 'static,
   S::Error: Into<BoxError> + Send + Sync + 'static,
   S::Future: Send + 'static,

Implementations

impl<S> Client<S>

pub fn new(connector: S) -> Self

Construct a new Client with a custom connector

pub fn with_retry_config(self, retry_config: RetryConfig) -> Self

impl Client<Standard>

pub fn https() -> StandardClient

Construct an https based client

impl<S> Client<S> where
S: Service<Request<SdkBody>, Response = Response<SdkBody>> + Send + Clone + 'static,
S::Error: Into<Box<dyn Error + Send + Sync>> + Send + Sync + 'static,
S::Future: Send + 'static,

pub async fn call<O, T, E, Retry>(
&self,
input: Operation<O, Retry>
) -> Result<T, SdkError<E>> where
O: ParseHttpResponse<SdkBody, Output = Result<T, E>> + Send + Sync + Clone + 'static,
E: Error + ProvideErrorKind,
Retry: ClassifyResponse<SdkSuccess<T>, SdkError<E>>,

Dispatch this request to the network

For ergonomics, this does not include the raw response for successful responses. To access the raw response use call_raw.

pub async fn call_raw<O, R, E, Retry>(
&self,
input: Operation<O, Retry>
) -> Result<SdkSuccess<R>, SdkError<E>> where
O: ParseHttpResponse<SdkBody, Output = Result<R, E>> + Send + Sync + Clone + 'static,
E: Error + ProvideErrorKind,
Retry: ClassifyResponse<SdkSuccess<R>, SdkError<E>>,

Dispatch this request to the network

The returned result contains the raw HTTP response which can be useful for debugging or implementing unsupported features.