Struct aws_hyper::DynConnector

#[non_exhaustive]
pub struct DynConnector(_);

A Smithy connector that uses dynamic dispatch.

This type allows you to pay a small runtime cost to avoid having to name the exact connector you’re using anywhere you want to hold a Client. Specifically, this will use Box to enable dynamic dispatch for every request that goes through the connector, which increases memory pressure and suffers an additional vtable indirection for each request, but is unlikely to matter in all but the highest-performance settings.

Implementations

impl DynConnector

pub fn new<E, C>(connector: C) -> DynConnector where
C: SmithyConnector<Error = E> + Send + 'static,
E: Into<Box<dyn Error + Sync + Send + 'static, Global>>,

Construct a new dynamically-dispatched Smithy middleware.